refactor(syn-docs): eliminate 14 fitness violations (ISS-283)
Extract monolithic components into focused modules to bring all metrics
under threshold. Removes all syn-docs entries from fitness-exceptions.toml.

Extractions:
- TopologyDependencyGraph: canvas-utils, useForceSimulation hook,
  TopologyTooltip, TopologyLegend
- Route files: lib/mdx-text, lib/mdx-files, lib/openapi-renderer,
  lib/openapi-sections (shared + decomposed helpers)
- HeroScene: hero/NetworkNodes (with pure geometry fn), hero/constants
- Pages: (home)/sections, llms/sections

Closes #283
