+++ .topology/manifest.toml
@@ -1,13 +1,13 @@
 [topology]
 version = "0.1.0"
-generated_at = "2025-12-17T18:05:21Z"
+generated_at = "2025-12-17T22:21:27Z"
 generator = "aps-cli"
 generator_version = "0.1.0"
 
 [analysis]
 root = "."
-languages = ["tsx", "typescript", "python", "rust"]
-total_files = 1197
-total_functions = 10208
-total_modules = 855
-total_dependencies = 4712
+languages = ["tsx", "python", "typescript", "rust"]
+total_files = 881
+total_functions = 7134
+total_modules = 595
+total_dependencies = 182
