+++ apps/syn-docs/app/llms-full.txt/route.ts
@@ -1,45 +1,8 @@
-import fs from 'node:fs';
 import path from 'node:path';
+import { collectMdxFiles } from '@/lib/mdx-files';
+import { stripFrontmatter, stripJSX, parseFrontmatter } from '@/lib/mdx-text';
 
 export const revalidate = false;
-
-function stripFrontmatter(content: string): string {
-  const match = content.match(/^---\n[\s\S]*?\n---\n/);
-  return match ? content.slice(match[0].length).trim() : content.trim();
-}
-
-function stripJSXComponents(content: string): string {
-  // Remove self-closing JSX tags like <SystemArchitectureDiagram />
-  return content.replace(/<[A-Z][A-Za-z]+ \/>/g, '').replace(/\n{3,}/g, '\n\n');
-}
-
-function getFrontmatter(content: string): { title?: string; description?: string } {
-  const match = content.match(/^---\n([\s\S]*?)\n---/);
-  if (!match) return {};
-  const fm: Record<string, string> = {};
-  for (const line of match[1].split('\n')) {
-    const [key, ...rest] = line.split(':');
-    if (key && rest.length) fm[key.trim()] = rest.join(':').trim();
-  }
-  return fm;
-}
-
-function collectMdxFiles(dir: string): { filepath: string; content: string }[] {
-  const results: { filepath: string; content: string }[] = [];
-
-  function walk(d: string) {
-    for (const entry of fs.readdirSync(d, { withFileTypes: true })) {
-      const full = path.join(d, entry.name);
-      if (entry.isDirectory()) walk(full);
-      else if (entry.name.endsWith('.mdx')) {
-        results.push({ filepath: full, content: fs.readFileSync(full, 'utf-8') });
-      }
-    }
-  }
-
-  walk(dir);
-  return results;
-}
 
 export function GET() {
   const contentDir = path.join(process.cwd(), 'content/docs');
@@ -53,7 +16,6 @@ export function GET() {
   sections.push('---');
   sections.push('');
 
-  // Sort: guide pages first, then api, then cli
   const order = ['index.mdx', 'guide/', 'api/', 'cli/'];
   files.sort((a, b) => {
     const relA = path.relative(contentDir, a.filepath);
@@ -64,8 +26,8 @@ export function GET() {
   });
 
   for (const file of files) {
-    const fm = getFrontmatter(file.content);
-    const body = stripJSXComponents(stripFrontmatter(file.content));
+    const fm = parseFrontmatter(file.content);
+    const body = stripJSX(stripFrontmatter(file.content));
     const rel = path.relative(contentDir, file.filepath);
 
     if (!body) continue;
